#909 change: extract consent management
